Decoupled (in a way) Projector from DataKeeper + Projector has new methods doing the same but returning tuple instead of modifying referenced variables + DataKeeper signal requesting shared_ptr<Projector> #23 almost there
Note: there was an attempt to fully decouple Projector from DataKeeper, however wire-frame Globe is calculated in DataKeeper which requires a lot of calls to Projector. For instance DataKeeper->composeWireGlobe() makes around 24k calls to Projector. And normal FPS is 59-60 which makes a lot of call per second. When it's done via boost.signal2 the view freezes severely. This was not acceptable, so the decision is to create shared_ptr<Projector> in GlobeViewer and sent a copy to DataKeeper via signal before actual work.