GLFW example full screen mode + press F11 to toggle full screen mode
Note: although it works, performance becomes unbearably slow. I don't know the reason for this behavior. It's definitely not the window size; make a simple test: being in windowed mode expand the window to the maximum, it will cover the whole screen and be fast, once you press F11 it'll become at least twice slower.